Final frontend changes, cleaned up endpoint
